+++ CHANGELOG.md
@@ -4,6 +4,7 @@ ThreePhasePowerModels.jl Change Log
 ### staged
 - Allow for arbitrarily named sourcebus
 - Add json parser
+- Fix bug in OpenDSS parse of Capacitors [zbase factor and wrong sign] (#138)
 
 ### v0.3.0
 - Update to JuMP v0.19/MathOptInterface